requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
python-dotenv==1.1.1
python-telegram-bot==22.5
reportlab==4.4.4
Pillow==11.3.0
PyPDF2==3.0.1
pdfplumber==0.11.7
python-docx==1.2.0
httpx==0.28.1
pdfminer.six==20250506
